+++ .github/workflows/node-ci.yml
@@ -1,0 +1,43 @@
+name: Node.js CI
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  ci:
+    name: Lint / Build / Test
+    runs-on: ubuntu-20.04
+
+    strategy:
+      matrix:
+        node-version: [18.x]
+
+    steps:
+      - name: Checkout Source Files
+        uses: actions/checkout@8e5e7e5ab8b370d6c329ec480221332ada57f0ab # v3
+
+      - name: Use Node.js v${{ matrix.node-version }}
+        uses: actions/setup-node@64ed1c7eab4cce3362f8c340dee64e5eaeef8f7c # v3
+        with:
+          node-version: ${{ matrix.node-version }}
+
+      - name: Install Dependencies
+        run: npm ci
+
+      - name: Check for eslint-disable
+        uses: naomi-lgbt/eslint-disable-action@main
+        with:
+          directories: ["src"]
+
+      - name: Lint Source Files
+        run: npm run lint
+
+      - name: Verify Build
+        run: npm run build
+
+      - name: Run Tests
+        run: npm run test
