+++ action.yml
@@ -6,4 +6,4 @@ inputs:
     required: true
 runs:
   using: 'node16'
-  main: 'index.js'
+  main: 'dist/index.js'
